// Qanat.Web/src/app/shared/generated/model/support-ticket-note-simple-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class SupportTicketNoteSimpleDto { 
    SupportTicketNoteID?: number;
    SupportTicketID?: number;
    Message?: string | null;
    InternalNote?: boolean;
    CreateUserID?: number;
    CreateDate?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface SupportTicketNoteSimpleDtoForm { 
    SupportTicketNoteID?: FormControl<number>;
    SupportTicketID?: FormControl<number>;
    Message?: FormControl<string>;
    InternalNote?: FormControl<boolean>;
    CreateUserID?: FormControl<number>;
    CreateDate?: FormControl<string>;
}

export class SupportTicketNoteSimpleDtoFormControls { 
    public static SupportTicketNoteID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SupportTicketID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Message = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static InternalNote = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CreateUserID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CreateDate = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}